# backend/api.yaml
openapi: 3.0.3
info:
  title: Budget App API
  version: 1.0.0
  description: API para presupuestos con ciclos automáticos, provisiones de gastos, gastos recurrentes y gastos manuales.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Refrescar token de acceso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        "200":
          description: Token refrescado
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
...